requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.0rc9
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.2
outcome==1.2.0
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.4
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.1
zipp==3.9.0
